refactor the functions into ./src
